Change to use QRGenerator
