+++ Trust/Extensions/UIImage.swift
@@ -17,18 +17,4 @@ extension UIImage {
         UIGraphicsEndImageContext()
         return image!
     }
-
-    func generateQRCode(from string: String) -> UIImage? {
-        let context = CIContext()
-        let data = string.data(using: String.Encoding.ascii)
-
-        if let filter = CIFilter(name: "CIQRCodeGenerator") {
-            filter.setValue(data, forKey: "inputMessage")
-            let transform = CGAffineTransform(scaleX: 7, y: 7)
-            if let output = filter.outputImage?.transformed(by: transform), let cgImage = context.createCGImage(output, from: output.extent) {
-                return UIImage(cgImage: cgImage)
-            }
-        }
-        return nil
-    }
 }
